Autonomous firmare
Updated firmware to version 1.1 featuring autonomous operation of the
robot.  If the robot encounters an obstacle ahead, it randomly chooses a
direction to pivot.  The robot pivots in that direction and measures
distance ahead until either finds an open pathway to proceed forward or
else it gives up ("stuck") and defaults back to the manual mode.  Readme
has been updated accordingly.