Create Photon_Robot_US-test.ino
This test routine tests out three HC-SR04 ultrasound modules whose triggers are via the shift register and echo pins go to 5v tolerant inputs on the Photon.  A new reading is taken from a sensor and published to the Particle cloud every 1 seconds.  The Particle console is used to display the results for testing.